Added the vauTexEdit project from Eclipse
This project allows for viewing and editing of .tex files and also
allows files that use the vaucanson-g to be compiled.
NOTE: This project does not contain any LaTex executables, but uses the
latex, dvips, and pstopdf compilers that are already installed on the
users system.
Also, since system calls must be made to call these compilers into
action this particular version of VauTexEdit is only compatible with OS
X.  I would hope that other users will improve the compitibility issues
inherent with this design very soon.