Added line numbering and began preparation to support windows.
Added line numbering and simplified the compile method in TexEdtiViewer
class to remove system dependant syntax. (in this case it was a '/').
All system dependant sysntax will be isolated to the LateCompile class
which will handle this with a System.getProperty("os.name") call to get
the operating system's name and an if statement to select the correct
commands and sysntax.